requirements.txt
python-telegram-bot==20.3
python-dotenv
fastapi
uvicorn
whisper
torch